@@ -1,6 +1,6 @@
 [workspace.package]
 authors = ["Band Protocol <dev@bandprotocol.com>"]
-edition = "2021"
+edition = "2024"
 license = "MIT OR Apache-2.0"
 repository = "https://github.com/bandprotocol/bothan"
 exclude = [".github/"]
@@ -10,31 +10,40 @@ members = [
     "bothan-api/client/rust-client",
     "bothan-api/server",
     "bothan-api/server-cli",
-    "bothan-binance",
-    "bothan-bybit",
-    "bothan-coinbase",
-    "bothan-coingecko",
-    "bothan-coinmarketcap",
-    "bothan-core",
-    "bothan-cryptocompare",
-    "bothan-htx",
-    "bothan-kraken",
-    "bothan-okx",
+    "bothan-*",
 ]
 exclude = ["bothan-api", "bothan-api-proxy"]
 resolver = "2"
 
 [workspace.dependencies]
+bothan-api = { path = "bothan-api/server" }
+bothan-core = { path = "bothan-core" }
+bothan-client = { path = "bothan-api/client/rust-client" }
+bothan-lib = { path = "bothan-lib" }
+
+bothan-binance = { path = "bothan-binance" }
+bothan-bitfinex = { path = "bothan-bitfinex" }
+bothan-bybit = { path = "bothan-bybit" }
+bothan-coinbase = { path = "bothan-coinbase" }
+bothan-coingecko = { path = "bothan-coingecko" }
+bothan-coinmarketcap = { path = "bothan-coinmarketcap" }
+bothan-htx = { path = "bothan-htx" }
+bothan-kraken = { path = "bothan-kraken" }
+bothan-okx = { path = "bothan-okx" }
+
 anyhow = "1.0.86"
 async-trait = "0.1.77"
-chrono = "0.4.38"
-derive_more = { version = "1.0.0-beta.6", features = ["full"] }
+bincode = "2.0.0-rc.3"
+chrono = "0.4.39"
+derive_more = { version = "2.0.1", features = ["full"] }
+dirs = "6.0.0"
 futures = "0.3.30"
 futures-util = "0.3.31"
 humantime-serde = "1.1.1"
-itertools = "0.13.0"
+itertools = "0.14.0"
 mockito = "1.4.0"
-opentelemetry = { version = "0.26.0", features = ["metrics"] }
+num-traits = "0.2.19"
+opentelemetry = { version = "0.28.0", features = ["metrics"] }
 prost = "0.13.1"
 protoc-gen-prost = "0.4.0"
 protoc-gen-tonic = "0.4.1"
@@ -43,26 +52,11 @@ reqwest = { version = "0.12.3", features = ["json"] }
 rust_decimal = "1.10.2"
 semver = "1.0.23"
 serde = { version = "1.0.197", features = ["std", "derive", "alloc"] }
-serde_json = "1.0.116"
-thiserror = "2.0.6"
+serde_json = "1.0.138"
+thiserror = "2.0.11"
 tokio = { version = "1.36.0", features = ["full"] }
 tokio-tungstenite = { version = "0.24.0", features = ["native-tls"] }
 tonic = "0.12.1"
 tracing = "0.1.40"
 tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
 url = "2.5.0"
-
-bothan-core = { path = "bothan-core" }
-
-bothan-client = { path = "bothan-api/client/rust-client" }
-bothan-api = { path = "bothan-api/server" }
-
-bothan-binance = { path = "bothan-binance" }
-bothan-bybit = { path = "bothan-bybit" }
-bothan-coinbase = { path = "bothan-coinbase" }
-bothan-coingecko = { path = "bothan-coingecko" }
-bothan-coinmarketcap = { path = "bothan-coinmarketcap" }
-bothan-cryptocompare = { path = "bothan-cryptocompare" }
-bothan-htx = { path = "bothan-htx" }
-bothan-kraken = { path = "bothan-kraken" }
-bothan-okx = { path = "bothan-okx" }
