@@ -1,39 +1,33 @@
 use std::collections::HashMap;
-use std::fs::{create_dir_all, read_to_string, remove_dir_all, write, File};
+use std::fs::{File, create_dir_all, read_to_string, write};
 use std::io::BufReader;
 use std::path::PathBuf;
 use std::str::FromStr;
 use std::sync::Arc;
 use std::time::Duration;
 
-use anyhow::Context;
+use anyhow::{Context, bail};
+use bothan_api::api::BothanServer;
+use bothan_api::config::AppConfig;
+use bothan_api::config::ipfs::IpfsAuthentication;
+use bothan_api::config::manager::crypto_info::sources::CryptoSourceConfigs;
+use bothan_api::proto::bothan::v1::BothanServiceServer;
+use bothan_api::{REGISTRY_REQUIREMENT, VERSION};
+use bothan_core::ipfs::{IpfsClient, IpfsClientBuilder};
+use bothan_core::manager::CryptoAssetInfoManager;
+use bothan_core::manager::crypto_asset_info::worker::CryptoAssetWorker;
+use bothan_core::manager::crypto_asset_info::worker::opts::CryptoAssetWorkerOpts;
+use bothan_core::monitoring::{Client as MonitoringClient, Signer};
+use bothan_core::store::rocksdb::RocksDbStore;
+use bothan_lib::registry::{Registry, Valid};
+use bothan_lib::store::{RegistryStore, Store};
+use bothan_lib::worker::AssetWorker;
+use bothan_lib::worker::error::AssetWorkerError;
 use clap::Parser;
 use reqwest::header::{HeaderName, HeaderValue};
 use semver::{Version, VersionReq};
 use tonic::transport::Server;
 use tracing::{debug, error, info};
-
-use bothan_api::api::BothanServer;
-use bothan_api::config::ipfs::IpfsAuthentication;
-use bothan_api::config::manager::crypto_info::sources::CryptoSourceConfigs;
-use bothan_api::config::AppConfig;
-use bothan_api::proto::bothan::v1::BothanServiceServer;
-use bothan_api::{REGISTRY_REQUIREMENT, VERSION};
-use bothan_binance::BinanceWorkerBuilder;
-use bothan_bybit::BybitWorkerBuilder;
-use bothan_coinbase::CoinbaseWorkerBuilder;
-use bothan_coingecko::CoinGeckoWorkerBuilder;
-use bothan_coinmarketcap::CoinMarketCapWorkerBuilder;
-use bothan_core::ipfs::{IpfsClient, IpfsClientBuilder};
-use bothan_core::manager::CryptoAssetInfoManager;
-use bothan_core::monitoring::{Client as MonitoringClient, Signer};
-use bothan_core::registry::{Registry, Valid};
-use bothan_core::store::SharedStore;
-use bothan_core::worker::{AssetWorker, AssetWorkerBuilder};
-use bothan_cryptocompare::CryptoCompareWorkerBuilder;
-use bothan_htx::HtxWorkerBuilder;
-use bothan_kraken::KrakenWorkerBuilder;
-use bothan_okx::OkxWorkerBuilder;
 
 #[derive(Parser)]
 pub struct StartCli {
@@ -60,16 +54,14 @@ impl StartCli {
             Some(p) => {
                 let file = File::open(p).with_context(|| "Failed to open registry file")?;
                 let reader = BufReader::new(file);
-                serde_json::from_reader(reader).with_context(|| "Failed to parse registry")?
+                let registry =
+                    serde_json::from_reader(reader).with_context(|| "Failed to parse registry")?;
+                Some(registry)
             }
-            None => Registry::default(),
+            None => None,
         };
 
-        let valid_registry = registry
-            .validate()
-            .with_context(|| "Failed to validate registry")?;
-
-        let store = init_store(&app_config, valid_registry, self.unsafe_reset).await?;
+        let store = init_rocks_db_store(&app_config, registry, self.unsafe_reset).await?;
         let ipfs_client = init_ipfs_client(&app_config).await?;
         let monitoring_client = init_monitoring_client(&app_config).await?;
 
@@ -86,30 +78,38 @@ impl StartCli {
     }
 }
 
-async fn init_store(
+async fn init_rocks_db_store(
     config: &AppConfig,
-    registry: Registry<Valid>,
+    registry: Option<Registry<Valid>>,
     reset: bool,
-) -> anyhow::Result<SharedStore> {
-    if reset {
-        remove_dir_all(&config.store.path).with_context(|| "Failed to remove store directory")?;
-    }
+) -> anyhow::Result<RocksDbStore> {
+    let flush_path = &config.store.path;
 
-    if !config.store.path.is_dir() {
-        create_dir_all(&config.store.path).with_context(|| "Failed to create home directory")?;
-    }
+    let store = match (reset, flush_path.is_dir()) {
+        // If reset is true and the path is a directory, remove the directory and create a new store
+        (true, true) => {
+            let db = RocksDbStore::new(flush_path)?;
+            debug!("store reset successfully at {:?}", &flush_path);
+            db
+        }
+        // If no reset, load the store
+        (false, true) => {
+            let db = RocksDbStore::load(flush_path)?;
+            debug!("store loaded successfully at {:?}", &flush_path);
+            db
+        }
+        // If the path does not exist, create the directory and create a new store
+        (_, false) => {
+            create_dir_all(flush_path).with_context(|| "Failed to create home directory")?;
+            let db = RocksDbStore::new(flush_path)?;
+            debug!("store created successfully at {:?}", &flush_path);
+            db
+        }
+    };
 
-    let mut store = SharedStore::new(registry, &config.store.path)
-        .await
-        .with_context(|| "Failed to create store")?;
-
-    debug!("store created successfully at {:?}", &config.store.path);
-
-    if !reset {
-        store
-            .restore()
-            .await
-            .with_context(|| "Failed to restore store state")?;
+    // If a registry is provided, overwrite the registry
+    if let Some(registry) = registry {
+        store.set_registry(registry, "".to_string()).await?;
     }
 
     Ok(store)
@@ -165,13 +165,13 @@ async fn init_ipfs_client(config: &AppConfig) -> anyhow::Result<IpfsClient> {
     Ok(ipfs_client)
 }
 
-async fn init_bothan_server(
+async fn init_bothan_server<S: Store + 'static>(
     config: &AppConfig,
-    store: SharedStore,
+    store: S,
     ipfs_client: IpfsClient,
     monitoring_client: Option<Arc<MonitoringClient>>,
-) -> anyhow::Result<Arc<BothanServer>> {
-    let manager_store = SharedStore::create_manager_store(&store);
+) -> anyhow::Result<Arc<BothanServer<S>>> {
+    let manager_store = RegistryStore::new(store.clone());
 
     let stale_threshold = config.manager.crypto.stale_threshold;
     let bothan_version =
@@ -179,7 +179,12 @@ async fn init_bothan_server(
     let registry_version_requirement = VersionReq::from_str(REGISTRY_REQUIREMENT)
         .with_context(|| "Failed to parse registry version requirement")?;
 
-    let workers = init_crypto_workers(&store, &config.manager.crypto.source).await?;
+    let workers = match init_crypto_workers(&store, &config.manager.crypto.source).await {
+        Ok(workers) => workers,
+        Err(e) => {
+            bail!("failed to initialize workers: {:?}", e);
+        }
+    };
     let manager = CryptoAssetInfoManager::new(
         workers,
         manager_store,
@@ -210,79 +215,40 @@ async fn init_bothan_server(
     Ok(Arc::new(BothanServer::new(manager)))
 }
 
-async fn init_crypto_workers(
-    store: &SharedStore,
+async fn init_crypto_workers<S: Store + 'static>(
+    store: &S,
     source: &CryptoSourceConfigs,
-) -> anyhow::Result<HashMap<String, Arc<dyn AssetWorker>>> {
-    type Binance = BinanceWorkerBuilder;
-    type Bybit = BybitWorkerBuilder;
-    type Coinbase = CoinbaseWorkerBuilder;
-    type CoinGecko = CoinGeckoWorkerBuilder;
-    type CoinMarketCap = CoinMarketCapWorkerBuilder;
-    type CryptoCompare = CryptoCompareWorkerBuilder;
-    type Htx = HtxWorkerBuilder;
-    type Kraken = KrakenWorkerBuilder;
-    type Okx = OkxWorkerBuilder;
-
+) -> Result<HashMap<String, CryptoAssetWorker<S>>, AssetWorkerError> {
     let mut workers = HashMap::new();
 
-    if let Some(opts) = &source.binance {
-        add_worker::<Binance>(&mut workers, store, opts).await?;
-    }
-
-    if let Some(opts) = &source.bybit {
-        add_worker::<Bybit>(&mut workers, store, opts).await?;
-    }
-
-    if let Some(opts) = &source.coinbase {
-        add_worker::<Coinbase>(&mut workers, store, opts).await?;
-    }
-
-    if let Some(opts) = &source.coingecko {
-        add_worker::<CoinGecko>(&mut workers, store, opts).await?;
-    }
-
-    if let Some(opts) = &source.coinmarketcap {
-        add_worker::<CoinMarketCap>(&mut workers, store, opts).await?;
-    }
-
-    if let Some(opts) = &source.cryptocompare {
-        add_worker::<CryptoCompare>(&mut workers, store, opts).await?;
-    }
-
-    if let Some(opts) = &source.htx {
-        add_worker::<Htx>(&mut workers, store, opts).await?;
-    }
-
-    if let Some(opts) = &source.kraken {
-        add_worker::<Kraken>(&mut workers, store, opts).await?;
-    }
-
-    if let Some(opts) = &source.okx {
-        add_worker::<Okx>(&mut workers, store, opts).await?;
-    }
+    add_worker(&mut workers, store, &source.binance).await?;
+    add_worker(&mut workers, store, &source.bitfinex).await?;
+    add_worker(&mut workers, store, &source.bybit).await?;
+    add_worker(&mut workers, store, &source.coinbase).await?;
+    add_worker(&mut workers, store, &source.coingecko).await?;
+    add_worker(&mut workers, store, &source.coinmarketcap).await?;
+    add_worker(&mut workers, store, &source.htx).await?;
+    add_worker(&mut workers, store, &source.kraken).await?;
+    add_worker(&mut workers, store, &source.okx).await?;
 
     Ok(workers)
 }
 
-async fn add_worker<B>(
-    workers: &mut HashMap<String, Arc<dyn AssetWorker>>,
-    store: &SharedStore,
-    opts: &B::Opts,
-) -> anyhow::Result<()>
+async fn add_worker<S, O>(
+    workers: &mut HashMap<String, CryptoAssetWorker<S>>,
+    store: &S,
+    opts: &Option<O>,
+) -> Result<(), AssetWorkerError>
 where
-    B: AssetWorkerBuilder<'static>,
-    B::Error: Send + Sync + 'static,
-    B::Opts: Clone,
+    S: Store + 'static,
+    O: Clone + Into<CryptoAssetWorkerOpts>,
 {
-    let worker_name = B::worker_name();
-    let worker_store = SharedStore::create_worker_store(store, worker_name);
-    let worker = B::new(worker_store, opts.clone())
-        .build()
-        .await
-        .with_context(|| format!("Failed to build worker {worker_name}"))?;
+    if let Some(opts) = opts {
+        let worker = CryptoAssetWorker::build(opts.clone().into(), store).await?;
+        let worker_name = worker.name();
+        workers.insert(worker_name.to_string(), worker);
+        info!("loaded {} worker", worker_name);
+    }
 
-    workers.insert(worker_name.to_string(), worker);
-    info!("loaded {} worker", worker_name);
     Ok(())
 }
