@@ -1,12 +1,11 @@
 use std::path::PathBuf;
 use std::str::FromStr;
 
-use clap::{Parser, Subcommand};
-use tracing_subscriber::filter::Directive;
-use tracing_subscriber::EnvFilter;
-
-use bothan_api::config::log::LogLevel;
 use bothan_api::config::AppConfig;
+use bothan_api::config::log::LogLevel;
+use clap::{Parser, Subcommand};
+use tracing_subscriber::EnvFilter;
+use tracing_subscriber::filter::Directive;
 
 use crate::commands::config::ConfigCli;
 use crate::commands::key::KeyCli;
@@ -26,6 +25,9 @@ struct Cli {
     // global args
     #[arg(long, global = true)]
     config: Option<PathBuf>,
+
+    #[arg(long, short)]
+    version: bool,
 }
 
 #[derive(Subcommand)]
@@ -49,6 +51,11 @@ async fn main() {
             std::process::exit(1);
         }
     };
+
+    if cli.version {
+        println!("bothan-api {}", env!("CARGO_PKG_VERSION"));
+        std::process::exit(0);
+    }
 
     let config_path = &cli.config.unwrap_or(bothan_home_dir().join("config.toml"));
 
