@@ -1,11 +1,11 @@
 use std::sync::Arc;
 
+use bothan_core::manager::CryptoAssetInfoManager;
+use bothan_core::manager::crypto_asset_info::error::{PushMonitoringRecordError, SetRegistryError};
+use bothan_lib::store::Store;
 use semver::Version;
 use tonic::{Request, Response, Status};
 use tracing::{debug, error, info};
-
-use bothan_core::manager::crypto_asset_info::error::{PushMonitoringRecordError, SetRegistryError};
-use bothan_core::manager::CryptoAssetInfoManager;
 
 use crate::api::utils::parse_price_state;
 use crate::proto::bothan::v1::{
@@ -17,20 +17,20 @@ use crate::proto::bothan::v1::{
 pub const PRECISION: u32 = 9;
 
 /// The `BothanServer` struct represents a server that implements the `BothanService` trait.
-pub struct BothanServer {
-    manager: Arc<CryptoAssetInfoManager<'static>>,
+pub struct BothanServer<S: Store + 'static> {
+    manager: Arc<CryptoAssetInfoManager<S>>,
 }
 
-impl BothanServer {
+impl<S: Store> BothanServer<S> {
     /// Creates a new `CryptoQueryServer` instance.
-    pub fn new(manager: Arc<CryptoAssetInfoManager<'static>>) -> Self {
+    pub fn new(manager: Arc<CryptoAssetInfoManager<S>>) -> Self {
         BothanServer { manager }
     }
 }
 
 // TODO: cleanup logging with span
 #[tonic::async_trait]
-impl BothanService for BothanServer {
+impl<S: Store> BothanService for BothanServer<S> {
     async fn get_info(
         &self,
         _: Request<GetInfoRequest>,
@@ -83,19 +83,19 @@ impl BothanService for BothanServer {
                 Err(Status::invalid_argument("Registry is invalid"))
             }
             Err(SetRegistryError::UnsupportedVersion) => {
-                error!("invalid registry");
-                Err(Status::invalid_argument("Registry is invalid"))
+                error!("unsupported registry version");
+                Err(Status::invalid_argument("Registry version is unsupported"))
             }
             Err(SetRegistryError::FailedToParse) => {
                 error!("failed to parse registry");
-                Err(Status::invalid_argument("Registry is invalid"))
+                Err(Status::invalid_argument("Unable to parse registry version"))
             }
             Err(SetRegistryError::InvalidHash) => {
                 error!("invalid IPFS hash");
                 Err(Status::invalid_argument("Invalid IPFS hash"))
             }
-            Err(SetRegistryError::FailedToSetRegistry(e)) => {
-                error!("failed to set registry: {e}");
+            Err(SetRegistryError::FailedToSetRegistry) => {
+                error!("failed to set registry");
                 Err(Status::internal("Failed to set registry"))
             }
         }
