@@ -1,41 +1,31 @@
 use serde::{Deserialize, Serialize};
 
-use bothan_binance::BinanceWorkerBuilderOpts;
-use bothan_bybit::BybitWorkerBuilderOpts;
-use bothan_coinbase::CoinbaseWorkerBuilderOpts;
-use bothan_coingecko::CoinGeckoWorkerBuilderOpts;
-use bothan_coinmarketcap::CoinMarketCapWorkerBuilderOpts;
-use bothan_cryptocompare::CryptoCompareWorkerBuilderOpts;
-use bothan_htx::HtxWorkerBuilderOpts;
-use bothan_kraken::KrakenWorkerBuilderOpts;
-use bothan_okx::OkxWorkerBuilderOpts;
-
 /// The configuration for the worker sources.
 #[derive(Clone, Debug, Default, Serialize, Deserialize)]
 pub struct CryptoSourceConfigs {
-    pub binance: Option<BinanceWorkerBuilderOpts>,
-    pub bybit: Option<BybitWorkerBuilderOpts>,
-    pub coinbase: Option<CoinbaseWorkerBuilderOpts>,
-    pub coingecko: Option<CoinGeckoWorkerBuilderOpts>,
-    pub coinmarketcap: Option<CoinMarketCapWorkerBuilderOpts>,
-    pub cryptocompare: Option<CryptoCompareWorkerBuilderOpts>,
-    pub htx: Option<HtxWorkerBuilderOpts>,
-    pub kraken: Option<KrakenWorkerBuilderOpts>,
-    pub okx: Option<OkxWorkerBuilderOpts>,
+    pub binance: Option<bothan_binance::WorkerOpts>,
+    pub bitfinex: Option<bothan_bitfinex::WorkerOpts>,
+    pub bybit: Option<bothan_bybit::WorkerOpts>,
+    pub coinbase: Option<bothan_coinbase::WorkerOpts>,
+    pub coingecko: Option<bothan_coingecko::WorkerOpts>,
+    pub coinmarketcap: Option<bothan_coinmarketcap::WorkerOpts>,
+    pub htx: Option<bothan_htx::WorkerOpts>,
+    pub kraken: Option<bothan_kraken::WorkerOpts>,
+    pub okx: Option<bothan_okx::WorkerOpts>,
 }
 
 impl CryptoSourceConfigs {
     pub fn with_default_sources() -> Self {
         CryptoSourceConfigs {
-            binance: Some(BinanceWorkerBuilderOpts::default()),
-            bybit: Some(BybitWorkerBuilderOpts::default()),
-            coinbase: Some(CoinbaseWorkerBuilderOpts::default()),
-            coingecko: Some(CoinGeckoWorkerBuilderOpts::default()),
-            coinmarketcap: Some(CoinMarketCapWorkerBuilderOpts::default()),
-            cryptocompare: Some(CryptoCompareWorkerBuilderOpts::default()),
-            htx: Some(HtxWorkerBuilderOpts::default()),
-            kraken: Some(KrakenWorkerBuilderOpts::default()),
-            okx: Some(OkxWorkerBuilderOpts::default()),
+            binance: Some(bothan_binance::WorkerOpts::default()),
+            bitfinex: Some(bothan_bitfinex::WorkerOpts::default()),
+            bybit: Some(bothan_bybit::WorkerOpts::default()),
+            coinbase: Some(bothan_coinbase::WorkerOpts::default()),
+            coingecko: Some(bothan_coingecko::WorkerOpts::default()),
+            coinmarketcap: Some(bothan_coinmarketcap::WorkerOpts::default()),
+            htx: Some(bothan_htx::WorkerOpts::default()),
+            kraken: Some(bothan_kraken::WorkerOpts::default()),
+            okx: Some(bothan_okx::WorkerOpts::default()),
         }
     }
 }
