@@ -1,5 +1,5 @@
 pub use error::{ConnectionError, MessageError, SendError};
-pub use websocket::{BinanceWebSocketConnection, BinanceWebSocketConnector};
+pub use websocket::{WebSocketConnection, WebSocketConnector};
 
 pub mod error;
 pub mod msgs;
