@@ -1,5 +1,5 @@
 pub use error::{ConnectionError, MessageError, SendError};
-pub use websocket::{BybitWebSocketConnection, BybitWebSocketConnector};
+pub use websocket::{WebSocketConnection, WebSocketConnector};
 
 pub mod error;
 pub mod types;
