@@ -6,10 +6,10 @@ use crate::worker::types::DEFAULT_CHANNEL_SIZE;
 /// Options for configuring the `BybitWorkerBuilder`.
 ///
 /// `BybitWorkerBuilderOpts` provides a way to specify custom settings for creating a `BybitWorker`.
-/// This struct allows users to set optional parameters such as the WebSocket URL and the internal channel size,
-/// which will be used during the construction of the `BybitWorker`.
+/// This struct allows users to set optional parameters such as the WebSocket URL and the internal
+/// channel size, which will be used during the construction of the `BybitWorker`.
 #[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct BybitWorkerBuilderOpts {
+pub struct WorkerOpts {
     #[serde(default = "default_url")]
     pub url: String,
     #[serde(default = "default_internal_ch_size")]
@@ -24,7 +24,7 @@ fn default_internal_ch_size() -> usize {
     DEFAULT_CHANNEL_SIZE
 }
 
-impl Default for BybitWorkerBuilderOpts {
+impl Default for WorkerOpts {
     fn default() -> Self {
         Self {
             url: default_url(),
