@@ -1,6 +1,6 @@
 pub use error::{ConnectionError, MessageError, SendError};
 pub use types::channels::ticker::Ticker;
-pub use websocket::{CoinbaseWebSocketConnection, CoinbaseWebSocketConnector};
+pub use websocket::{WebSocketConnection, WebSocketConnector};
 
 pub mod error;
 pub mod types;
