@@ -1,8 +1,6 @@
-pub use api::websocket::{CoinbaseWebSocketConnection, CoinbaseWebSocketConnector};
-pub use worker::builder::CoinbaseWorkerBuilder;
-pub use worker::error::BuildError;
-pub use worker::opts::CoinbaseWorkerBuilderOpts;
-pub use worker::CoinbaseWorker;
+pub use api::websocket::{WebSocketConnection, WebSocketConnector};
+pub use worker::Worker;
+pub use worker::opts::WorkerOpts;
 
 pub mod api;
 pub mod worker;
