@@ -9,7 +9,7 @@ pub(crate) const DEFAULT_PRO_URL: &str = "https://pro-api.coingecko.com/api/v3/"
 pub(crate) const API_KEY_HEADER: &str = "x-cg-pro-api-key";
 
 /// Represents a coin with basic information.
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Coin {
     pub id: String,
     pub symbol: String,
@@ -17,7 +17,7 @@ pub struct Coin {
 }
 
 /// Represents market data for a coin.
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Price {
     pub usd: f64,
     pub last_updated_at: i64,
