@@ -1,41 +1,59 @@
-use bothan_core::store::error::Error as StoreError;
-use bothan_core::store::WorkerStore;
-use bothan_core::worker::{AssetState, AssetWorker, SetQueryIDError};
+use std::collections::HashSet;
+use std::sync::{Arc, Weak};
 
-use crate::api::CoinGeckoRestAPI;
+use bothan_lib::store::{Store, WorkerStore};
+use bothan_lib::types::AssetState;
+use bothan_lib::worker::AssetWorker;
+use bothan_lib::worker::error::AssetWorkerError;
+use bothan_lib::worker::rest::{AssetInfoProvider, start_polling};
 
-mod asset_worker;
-pub mod builder;
-pub mod error;
+use crate::WorkerOpts;
+use crate::api::error::ProviderError;
+use crate::api::{RestApi, RestApiBuilder};
+
 pub mod opts;
-pub mod types;
 
-/// A worker that fetches and stores the asset information from CoinGecko's API.
-pub struct CoinGeckoWorker {
-    api: CoinGeckoRestAPI,
-    store: WorkerStore,
-}
+const WORKER_NAME: &str = "coingecko";
 
-impl CoinGeckoWorker {
-    /// Create a new worker with the specified api and store.
-    pub fn new(api: CoinGeckoRestAPI, store: WorkerStore) -> Self {
-        Self { api, store }
-    }
+pub struct Worker<S: Store> {
+    // The `api` is owned by this struct to ensure that any weak references
+    // are properly cleaned up when the worker is dropped.
+    #[allow(dead_code)]
+    api: Arc<RestApi>,
+    store: WorkerStore<S>,
 }
 
 #[async_trait::async_trait]
-impl AssetWorker for CoinGeckoWorker {
-    /// Fetches the AssetStatus for the given cryptocurrency ids.
-    async fn get_asset(&self, id: &str) -> Result<AssetState, StoreError> {
-        self.store.get_asset(&id).await
+impl<S: Store + 'static> AssetWorker<S> for Worker<S> {
+    type Opts = WorkerOpts;
+
+    fn name(&self) -> &'static str {
+        WORKER_NAME
     }
 
-    /// Adds the specified cryptocurrency IDs to the query set.
-    async fn set_query_ids(&self, ids: Vec<String>) -> Result<(), SetQueryIDError> {
-        self.store
-            .set_query_ids(ids)
-            .await
-            .map_err(|e| SetQueryIDError::new(e.to_string()))?;
+    async fn build(opts: Self::Opts, store: &S) -> Result<Self, AssetWorkerError> {
+        let api = Arc::new(RestApiBuilder::new(opts.url, opts.user_agent, opts.api_key).build()?);
+
+        let worker_store = WorkerStore::new(store, WORKER_NAME);
+
+        tokio::spawn(start_polling(
+            opts.update_interval,
+            Arc::downgrade(&api) as Weak<dyn AssetInfoProvider<Error = ProviderError>>,
+            worker_store.clone(),
+        ));
+
+        Ok(Worker {
+            api,
+            store: worker_store,
+        })
+    }
+
+    async fn get_asset(&self, id: &str) -> Result<AssetState, AssetWorkerError> {
+        Ok(self.store.get_asset(id).await?)
+    }
+
+    async fn set_query_ids(&self, ids: HashSet<String>) -> Result<(), AssetWorkerError> {
+        self.store.set_query_ids(ids).await?;
         Ok(())
     }
 }
