@@ -3,17 +3,19 @@ use std::time::Duration;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
 use crate::api::types::{DEFAULT_URL, DEFAULT_USER_AGENT};
-use crate::worker::types::DEFAULT_UPDATE_INTERVAL;
+
+const DEFAULT_UPDATE_INTERVAL: Duration = Duration::from_secs(60);
 
 /// Options for configuring the `CoinGeckoWorkerBuilder`.
 ///
-/// `CoinGeckoWorkerBuilderOpts` provides a way to specify custom settings for creating a `CoinGeckoWorker`.
-/// This struct allows users to set optional parameters such as the WebSocket URL and the internal channel size,
-/// which will be used during the construction of the `CoinGeckoWorker`.
+/// `CoinGeckoWorkerBuilderOpts` provides a way to specify custom settings for creating a
+/// `CoinGeckoWorker`. This struct allows users to set optional parameters such as the WebSocket URL
+/// and the internal channel size, which will be used during the construction of the
+/// `CoinGeckoWorker`.
 #[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct CoinGeckoWorkerBuilderOpts {
-    #[serde(default = "default_url")]
-    pub url: String,
+pub struct WorkerOpts {
+    #[serde(serialize_with = "none_is_default_url")]
+    pub url: Option<String>,
     #[serde(default)]
     #[serde(deserialize_with = "empty_string_is_none")]
     #[serde(serialize_with = "none_is_empty_string")]
@@ -25,10 +27,6 @@ pub struct CoinGeckoWorkerBuilderOpts {
     pub update_interval: Duration,
 }
 
-fn default_url() -> String {
-    DEFAULT_URL.to_string()
-}
-
 fn default_user_agent() -> String {
     DEFAULT_USER_AGENT.to_string()
 }
@@ -37,10 +35,10 @@ fn default_update_interval() -> Duration {
     DEFAULT_UPDATE_INTERVAL
 }
 
-impl Default for CoinGeckoWorkerBuilderOpts {
+impl Default for WorkerOpts {
     fn default() -> Self {
         Self {
-            url: default_url(),
+            url: None,
             api_key: None,
             user_agent: default_user_agent(),
             update_interval: default_update_interval(),
@@ -62,5 +60,15 @@ fn none_is_empty_string<S: Serializer>(
     match value {
         Some(val) => serializer.serialize_str(val),
         None => serializer.serialize_str(""),
+    }
+}
+
+fn none_is_default_url<S: Serializer>(
+    value: &Option<String>,
+    serializer: S,
+) -> Result<S::Ok, S::Error> {
+    match value {
+        Some(val) => serializer.serialize_str(val),
+        None => serializer.serialize_str(DEFAULT_URL),
     }
 }
