@@ -1,27 +1,27 @@
-use reqwest::header::{HeaderMap, HeaderValue};
 use reqwest::ClientBuilder;
+use reqwest::header::{HeaderMap, HeaderValue};
 use url::Url;
 
+use crate::api::RestApi;
 use crate::api::error::BuildError;
 use crate::api::types::DEFAULT_URL;
-use crate::api::CoinMarketCapRestAPI;
 
 /// Builds a CoinMarketCapRestAPI with custom parameters.
 /// Methods can be chained to set the parameters and the
 /// `CoinMarketCapRestAPI` is constructed
-/// by calling the [`build`](CoinMarketCapRestAPIBuilder::build) method.
-pub struct CoinMarketCapRestAPIBuilder {
+/// by calling the [`build`](RestApiBuilder::build) method.
+pub struct RestApiBuilder {
     url: String,
     api_key: Option<String>,
 }
 
-impl CoinMarketCapRestAPIBuilder {
+impl RestApiBuilder {
     pub fn new<T, U>(url: T, api_key: Option<U>) -> Self
     where
         T: Into<String>,
         U: Into<String>,
     {
-        CoinMarketCapRestAPIBuilder {
+        RestApiBuilder {
             url: url.into(),
             api_key: api_key.map(Into::into),
         }
@@ -42,31 +42,28 @@ impl CoinMarketCapRestAPIBuilder {
     }
 
     /// Creates the configured `CoinMarketCapRestAPI`.
-    pub fn build(self) -> Result<CoinMarketCapRestAPI, BuildError> {
+    pub fn build(self) -> Result<RestApi, BuildError> {
         let mut headers = HeaderMap::new();
 
         let parsed_url = Url::parse(&self.url)?;
 
-        let key = match &self.api_key {
-            Some(key) => key,
-            None => return Err(BuildError::MissingAPIKey()),
-        };
+        let api_key = self.api_key.ok_or(BuildError::MissingAPIKey)?;
 
-        let mut val = HeaderValue::from_str(key)?;
+        let mut val = HeaderValue::from_str(&api_key)?;
         val.set_sensitive(true);
         headers.insert("X-CMC_PRO_API_KEY", val);
 
         let client = ClientBuilder::new().default_headers(headers).build()?;
 
-        Ok(CoinMarketCapRestAPI::new(parsed_url, client))
+        Ok(RestApi::new(parsed_url, client))
     }
 }
 
-impl Default for CoinMarketCapRestAPIBuilder {
+impl Default for RestApiBuilder {
     /// Creates a new `CoinMarketCapRestAPIBuilder` with the
     /// default URL and no API key.
     fn default() -> Self {
-        CoinMarketCapRestAPIBuilder {
+        RestApiBuilder {
             url: DEFAULT_URL.into(),
             api_key: None,
         }
