@@ -1,6 +1,5 @@
-pub use worker::builder::CoinMarketCapWorkerBuilder;
-pub use worker::opts::CoinMarketCapWorkerBuilderOpts;
-pub use worker::CoinMarketCapWorker;
+pub use worker::Worker;
+pub use worker::opts::WorkerOpts;
 
 pub mod api;
 pub mod worker;
