@@ -1,7 +1,8 @@
 pub use manager::CryptoAssetInfoManager;
 
 pub mod error;
-mod manager;
-mod price;
-mod signal_ids;
+pub mod manager;
+pub mod price;
+pub mod signal_ids;
 pub mod types;
+pub mod worker;
