@@ -5,10 +5,10 @@ use futures_util::stream::{SplitSink, SplitStream};
 use futures_util::{SinkExt, StreamExt};
 use serde_json::json;
 use tokio::net::TcpStream;
+use tokio_tungstenite::tungstenite::Message;
 use tokio_tungstenite::tungstenite::error::Error as TungsteniteError;
 use tokio_tungstenite::tungstenite::http::StatusCode;
-use tokio_tungstenite::tungstenite::Message;
-use tokio_tungstenite::{connect_async, MaybeTlsStream, WebSocketStream};
+use tokio_tungstenite::{MaybeTlsStream, WebSocketStream, connect_async};
 use tracing::warn;
 
 use crate::api::error::{ConnectionError, MessageError, SendError};
@@ -127,10 +127,11 @@ impl HtxWebSocketConnection {
 
 #[cfg(test)]
 pub(crate) mod test {
-    use super::*;
-    use crate::api::types::{DataUpdate, HtxResponse, Ping, SubResponse, Tick, UnsubResponse};
     use tokio::sync::mpsc;
     use ws_mock::ws_mock_server::{WsMock, WsMockServer};
+
+    use super::*;
+    use crate::api::types::{DataUpdate, HtxResponse, Ping, SubResponse, Tick, UnsubResponse};
 
     pub(crate) async fn setup_mock_server() -> WsMockServer {
         WsMockServer::start().await
