@@ -1,6 +1,5 @@
-pub use worker::builder::HtxWorkerBuilder;
-pub use worker::opts::HtxWorkerBuilderOpts;
-pub use worker::HtxWorker;
+pub use worker::Worker;
+pub use worker::opts::WorkerOpts;
 
 pub mod api;
 pub mod worker;
