@@ -1,16 +1,16 @@
 use futures_util::stream::{SplitSink, SplitStream};
 use futures_util::{SinkExt, StreamExt};
 use tokio::net::TcpStream;
+use tokio_tungstenite::tungstenite::Message;
 use tokio_tungstenite::tungstenite::error::Error as TungsteniteError;
 use tokio_tungstenite::tungstenite::http::StatusCode;
-use tokio_tungstenite::tungstenite::Message;
-use tokio_tungstenite::{connect_async, MaybeTlsStream, WebSocketStream};
+use tokio_tungstenite::{MaybeTlsStream, WebSocketStream, connect_async};
 use tracing::warn;
 
 use crate::api::error::{ConnectionError, MessageError, SendError};
+use crate::api::types::KrakenResponse;
 use crate::api::types::channel::ticker::{EventTrigger, TickerRequestParameters};
 use crate::api::types::message::{Method, PublicMessage};
-use crate::api::types::KrakenResponse;
 
 /// A connector for establishing a WebSocket connection to the Kraken API.
 pub struct KrakenWebSocketConnector {
@@ -139,9 +139,8 @@ pub(crate) mod test {
     use tokio::sync::mpsc;
     use ws_mock::ws_mock_server::{WsMock, WsMockServer};
 
-    use crate::api::types::{ChannelResponse, KrakenResponse, TickerResponse};
-
     use super::*;
+    use crate::api::types::{ChannelResponse, KrakenResponse, TickerResponse};
 
     pub(crate) async fn setup_mock_server() -> WsMockServer {
         WsMockServer::start().await
