@@ -1,5 +1,5 @@
 pub use error::{ConnectionError, MessageError, SendError};
-pub use websocket::{OkxWebSocketConnection, OkxWebSocketConnector};
+pub use websocket::{WebSocketConnection, WebsocketConnector};
 
 pub mod error;
 pub mod types;
