@@ -1,68 +1,88 @@
-use tokio::sync::mpsc::Sender;
+use std::collections::HashSet;
+use std::sync::Arc;
 
-use bothan_core::store::error::Error as StoreError;
-use bothan_core::store::WorkerStore;
-use bothan_core::worker::{AssetState, AssetWorker, SetQueryIDError};
+use bothan_lib::store::{Store, WorkerStore};
+use bothan_lib::types::AssetState;
+use bothan_lib::worker::AssetWorker;
+use bothan_lib::worker::error::AssetWorkerError;
+use tokio::sync::mpsc::{Sender, channel};
 
-use crate::api::websocket::OkxWebSocketConnector;
+use crate::WorkerOpts;
+use crate::api::websocket::WebsocketConnector;
+use crate::worker::asset_worker::start_asset_worker;
 
 mod asset_worker;
-pub mod builder;
-pub(crate) mod error;
 pub mod opts;
-mod types;
 
-/// A worker that fetches and stores the asset information from Okx's API.
-pub struct OkxWorker {
-    connector: OkxWebSocketConnector,
-    store: WorkerStore,
-    subscribe_tx: Sender<Vec<String>>,
-    unsubscribe_tx: Sender<Vec<String>>,
-}
+const WORKER_NAME: &str = "okx";
 
-impl OkxWorker {
-    /// Create a new worker with the specified connector, store and channels.
-    pub fn new(
-        connector: OkxWebSocketConnector,
-        store: WorkerStore,
-        subscribe_tx: Sender<Vec<String>>,
-        unsubscribe_tx: Sender<Vec<String>>,
-    ) -> Self {
-        Self {
-            connector,
-            store,
-            subscribe_tx,
-            unsubscribe_tx,
-        }
-    }
+pub struct Worker<S: Store> {
+    inner: Arc<InnerWorker<S>>,
 }
 
 #[async_trait::async_trait]
-impl AssetWorker for OkxWorker {
-    /// Fetches the AssetStatus for the given cryptocurrency id.
-    async fn get_asset(&self, id: &str) -> Result<AssetState, StoreError> {
-        self.store.get_asset(&id).await
+impl<S: Store + 'static> AssetWorker<S> for Worker<S> {
+    type Opts = WorkerOpts;
+
+    fn name(&self) -> &'static str {
+        WORKER_NAME
     }
 
-    /// Sets the specified cryptocurrency IDs to the query. If the ids are already in the query set,
-    /// it will not be resubscribed.
-    async fn set_query_ids(&self, ids: Vec<String>) -> Result<(), SetQueryIDError> {
-        let (to_sub, to_unsub) = self
-            .store
-            .set_query_ids(ids)
-            .await
-            .map_err(|e| SetQueryIDError::new(e.to_string()))?;
+    async fn build(opts: Self::Opts, store: &S) -> Result<Self, AssetWorkerError> {
+        let url = opts.url;
+        let ch_size = opts.internal_ch_size;
 
-        self.subscribe_tx
-            .send(to_sub)
-            .await
-            .map_err(|e| SetQueryIDError::new(e.to_string()))?;
+        let connector = WebsocketConnector::new(url);
+        let connection = connector.connect().await?;
 
-        self.unsubscribe_tx
-            .send(to_unsub)
-            .await
-            .map_err(|e| SetQueryIDError::new(e.to_string()))?;
+        let (sub_tx, sub_rx) = channel(ch_size);
+        let (unsub_tx, unsub_rx) = channel(ch_size);
+
+        let worker_store = WorkerStore::new(store, WORKER_NAME);
+        let to_sub = worker_store
+            .get_query_ids()
+            .await?
+            .into_iter()
+            .collect::<Vec<String>>();
+
+        if !to_sub.is_empty() {
+            sub_tx.send(to_sub).await?;
+        }
+
+        let inner = Arc::new(InnerWorker {
+            connector,
+            store: worker_store,
+            subscribe_tx: sub_tx,
+            unsubscribe_tx: unsub_tx,
+        });
+
+        tokio::spawn(start_asset_worker(
+            Arc::downgrade(&inner),
+            connection,
+            sub_rx,
+            unsub_rx,
+        ));
+
+        Ok(Worker { inner })
+    }
+
+    async fn get_asset(&self, id: &str) -> Result<AssetState, AssetWorkerError> {
+        Ok(self.inner.store.get_asset(id).await?)
+    }
+
+    async fn set_query_ids(&self, ids: HashSet<String>) -> Result<(), AssetWorkerError> {
+        let diff = self.inner.store.compute_query_id_difference(ids).await?;
+
+        self.inner.subscribe_tx.send(diff.added).await?;
+        self.inner.unsubscribe_tx.send(diff.removed).await?;
 
         Ok(())
     }
+}
+
+pub struct InnerWorker<S: Store> {
+    connector: WebsocketConnector,
+    store: WorkerStore<S>,
+    subscribe_tx: Sender<Vec<String>>,
+    unsubscribe_tx: Sender<Vec<String>>,
 }
