@@ -1,15 +1,11 @@
 use serde::{Deserialize, Serialize};
 
 use crate::api::types::DEFAULT_URL;
-use crate::worker::types::DEFAULT_CHANNEL_SIZE;
 
-/// Options for configuring the `OkxWorkerBuilder`.
-///
-/// `OkxWorkerBuilderOpts` provides a way to specify custom settings for creating a `OkxWorker`.
-/// This struct allows users to set optional parameters such as the WebSocket URL and the internal channel size,
-/// which will be used during the construction of the `OkxWorker`.
+pub const DEFAULT_CHANNEL_SIZE: usize = 1000;
+
 #[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct OkxWorkerBuilderOpts {
+pub struct WorkerOpts {
     #[serde(default = "default_url")]
     pub url: String,
     #[serde(default = "default_internal_ch_size")]
@@ -24,7 +20,7 @@ fn default_internal_ch_size() -> usize {
     DEFAULT_CHANNEL_SIZE
 }
 
-impl Default for OkxWorkerBuilderOpts {
+impl Default for WorkerOpts {
     fn default() -> Self {
         Self {
             url: default_url(),
